chore: move apollo to lens (#2261)
