+++ apps/prerender/package.json
@@ -12,9 +12,10 @@
     "lint:fix": "eslint . --fix --ext .js,.ts,.tsx"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.10",
     "axios": "^1.3.4",
+    "data": "*",
     "dotenv": "^16.0.3",
+    "lens": "*",
     "lib": "*",
     "next": "^13.2.4",
     "react": "^18.2.0",
@@ -25,7 +26,6 @@
   "devDependencies": {
     "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
-    "data": "*",
     "eslint-config-weblint": "*",
     "lens": "*",
     "tsconfig": "*",
