+++ apps/prerender/src/pages/u/[handle].tsx
@@ -1,8 +1,8 @@
 import Profile from '@components/Profile';
 import { HANDLE_SUFFIX, LENSPROTOCOL_HANDLE } from 'data/constants';
 import { PrerenderProfileDocument } from 'lens';
+import { nodeClient } from 'lens/apollo';
 import type { GetServerSidePropsContext } from 'next';
-import client from 'src/apollo';
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
   const handle = context.params?.handle;
@@ -22,7 +22,7 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
   } else {
     processedHandle = handle === LENSPROTOCOL_HANDLE ? handle : handle.concat(HANDLE_SUFFIX);
   }
-  const { data } = await client.query({
+  const { data } = await nodeClient.query({
     query: PrerenderProfileDocument,
     variables: { request: { handle: processedHandle } }
   });
