+++ apps/web/package.json
@@ -13,7 +13,6 @@
     "extract": "lingui compile; lingui extract"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.10",
     "@aws-sdk/client-s3": "^3.301.0",
     "@dnd-kit/core": "^6.0.7",
     "@dnd-kit/sortable": "^7.0.2",
@@ -38,6 +37,7 @@
     "@xmtp/xmtp-js": "^7.14.1",
     "axios": "^1.3.4",
     "clsx": "^1.2.1",
+    "data": "*",
     "dayjs": "^1.11.5",
     "dayjs-twitter": "^0.5.0",
     "dotenv": "^16.0.3",
@@ -45,6 +45,7 @@
     "framer-motion": "^10.9.2",
     "graphql": "^16.6.0",
     "jwt-decode": "^3.1.2",
+    "lens": "*",
     "lexical": "^0.9.1",
     "lib": "*",
     "mixpanel-browser": "^2.46.0",
@@ -83,9 +84,7 @@
     "@types/uuid": "^9.0.1",
     "abis": "*",
     "autoprefixer": "^10.4.14",
-    "data": "*",
     "eslint-config-weblint": "*",
-    "lens": "*",
     "postcss": "^8.4.21",
     "prettier-plugin-tailwindcss": "^0.2.5",
     "tailwindcss": "3.2.7",
