+++ apps/web/src/components/Publication/Actions/Mirror.tsx
@@ -1,4 +1,3 @@
-import type { ApolloCache } from '@apollo/client';
 import { SwitchHorizontalIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import onError from '@lib/onError';
@@ -15,9 +14,10 @@ import {
   useCreateMirrorTypedDataMutation,
   useCreateMirrorViaDispatcherMutation
 } from 'lens';
+import type { ApolloCache } from 'lens/apollo';
+import { publicationKeyFields } from 'lens/apollo/lib';
 import getSignature from 'lib/getSignature';
 import humanize from 'lib/humanize';
-import { publicationKeyFields } from 'lib/keyFields';
 import nFormatter from 'lib/nFormatter';
 import type { FC } from 'react';
 import { useState } from 'react';
