+++ apps/web/src/components/Publication/QueuedPublication.tsx
@@ -1,4 +1,3 @@
-import { useApolloClient } from '@apollo/client';
 import Attachments from '@components/Shared/Attachments';
 import IFramely from '@components/Shared/IFramely';
 import Markup from '@components/Shared/Markup';
@@ -11,6 +10,7 @@ import {
   useHasTxHashBeenIndexedQuery,
   usePublicationLazyQuery
 } from 'lens';
+import { useApolloClient } from 'lens/apollo';
 import getURLs from 'lib/getURLs';
 import type { FC } from 'react';
 import { useAppStore } from 'src/store/app';
