+++ packages/lens/package.json
@@ -9,7 +9,9 @@
     "typecheck": "tsc --pretty --noEmit"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.10"
+    "@apollo/client": "^3.7.10",
+    "axios": "^1.3.4",
+    "data": "*"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^3.2.2",
@@ -17,6 +19,7 @@
     "@graphql-codegen/typescript": "^3.0.2",
     "@graphql-codegen/typescript-operations": "^3.0.2",
     "@graphql-codegen/typescript-react-apollo": "^3.3.6",
+    "eslint-config-weblint": "*",
     "prettier": "^2.8.7",
     "tsconfig": "*",
     "typescript": "^5.0.2"
