+++ tests/package.json
@@ -11,6 +11,7 @@
     "data": "*",
     "eslint-config-weblint": "*",
     "lib": "*",
+    "lens": "*",
     "tsconfig": "*"
   }
 }
