+++ tests/scripts/packages/utils/parseJwt.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import parseJwt from 'lib/parseJwt';
+import { parseJwt } from 'lens/apollo/lib';
 
 test.describe('parseJwt', () => {
   test('should parse and return token expiration time', () => {
